Use only integer totals for easy/groove gauge recovery

+++ src/bms/player/beatoraja/play/GrooveGauge.java
@@ -263,7 +263,7 @@ public class GrooveGauge {
 			@Override
 			public float modify(float f, BMSModel model) {
 				if(f > 0) {
-					return (float) (f * model.getTotal() / model.getTotalNotes());
+					return (float) (f * (int)model.getTotal() / model.getTotalNotes());
 				}
 				return f;
 			}
